Tests for player view now working
