+++ sales-scenario/tests/functional/PlayerExtraFunctionsTest.php
@@ -4,11 +4,20 @@ use Illuminate\Foundation\Testing\WithoutMiddleware;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 
-class DashboardFuncTest extends TestCase
+class PartialViewListsTest extends TestCase
 {
     use WithoutMiddleware;
     use DatabaseTransactions;
 
+
+
+    private function getURL()
+    {
+        $this->login();
+        $expert = $this->getExpertWithPodcast();
+        $podcast = $expert->podcasts[0];
+        return $player = '/player/' . $expert->slug . '/' . $podcast->slug;
+    }
     public function test_expert_with_most_podcasts_show(){
 
         //Since the factory creates expert with max 10 podcasts, this expert must be most contributing
@@ -21,7 +30,7 @@ class DashboardFuncTest extends TestCase
 
         $mostContributing = $this->createExpertWithMultiplePodcasts(8);
 
-        $this->visit('dashboard')
+        $this->visit($this->getURL())
             ->see("$mostContributing->full_name")
             ->dontSee("$leastContributing->full_name");
     }
@@ -30,7 +39,7 @@ class DashboardFuncTest extends TestCase
 
         $expert_no_podcasts = factory(App\Expert::class)->create();
 
-        $this->visit('dashboard')
+        $this->visit($this->getURL())
             ->dontSeeLink("$expert_no_podcasts->full_name");
 
     }
@@ -45,7 +54,7 @@ class DashboardFuncTest extends TestCase
 
         //Tags are randomized, so lets make sure we see the 5 tags after a refresh
         for ($i = 0; $i < 2; $i++) {
-            $this->visit("dashboard");
+            $this->visit($this->getURL());
 
             foreach ($tags as $tag) {
                 $this->seeLink($tag->name);
@@ -59,7 +68,7 @@ class DashboardFuncTest extends TestCase
         }
 
         $expert_no_podcasts = factory(App\Expert::class)->create();
-        $this->visit('dashboard')
+        $this->visit($this->getURL())
             ->dontSeeLink("$expert_no_podcasts->full_name");
     }
 }
